+++ tests/links.spec.ts
@@ -113,8 +113,6 @@ test.describe('data', () => {
 
         const dump = await shouldBeDumpPage(page, 'get')
 
-        // TODO: Should this be in the query string? It's not, but... should it be? => See 'url' key in helpers.js
-
         await expect(dump.query).toEqual({ a: ['b', 'c'] })
         await expect(dump.method).toBe('get')
         await expect(dump.form).toEqual({})
@@ -135,10 +133,7 @@ test.describe('data', () => {
 
       const dump = await shouldBeDumpPage(page, 'get')
 
-      // TODO: Should this be in the query string? It's not, but... should it be? => See 'url' key in helpers.js
-      // const params = await page.evaluate(() => window.location.search)
-      // await expect(params).toBe('?foo=get')
-
+      await expect(page).toHaveURL('/dump/get?foo=get')
       await expect(dump.query).toEqual({ foo: 'get' })
       await expect(dump.method).toBe('get')
       await expect(dump.form).toEqual({})
@@ -339,7 +334,7 @@ test.describe('preserve state', () => {
       await expect(componentKey).not.toBeUndefined()
 
       await page.getByRole('link', { name: label }).click()
-      await expect(page).toHaveURL('/links/preserve-state-page-two')
+      await expect(page).toHaveURL(`/links/preserve-state-page-two?foo=${expected}`)
 
       const newComponentKey = await page.evaluate(() => (window as any)._inertia_page_key)
       await expect(newComponentKey).not.toBeUndefined()
@@ -373,7 +368,7 @@ test.describe('preserve state', () => {
       await expect(componentKey).not.toBeUndefined()
 
       await page.getByRole('link', { name: label }).click()
-      await expect(page).toHaveURL('/links/preserve-state-page-two')
+      await expect(page).toHaveURL(`/links/preserve-state-page-two?foo=${expected}`)
 
       // @ts-ignore
       const newComponentKey = await page.evaluate(() => window._inertia_page_key)
@@ -406,7 +401,7 @@ test.describe('preserve url', () => {
     const initialUrl = page.url()
 
     await page.getByRole('link', { name: '[URL] Preserve: false' }).click()
-    await expect(page).toHaveURL('/links/preserve-url-page-two')
+    await expect(page).toHaveURL('/links/preserve-url-page-two?foo=baz')
     await expect(page.getByText('Foo is now baz')).toBeVisible()
     await expect(page.url()).not.toBe(initialUrl) // URL should have changed
   })
@@ -455,7 +450,7 @@ test.describe('preserve scroll', () => {
 
   test('does not reset untracked scroll regions in persistent layouts', async ({ page }) => {
     await page.getByRole('link', { exact: true, name: 'Reset Scroll' }).click()
-    await expect(page).toHaveURL('/links/preserve-scroll-false-page-two')
+    await expect(page).toHaveURL('/links/preserve-scroll-false-page-two?foo=bar')
     await expect(page.getByText('Foo is now bar')).toBeVisible()
     await page.getByRole('button', { exact: true, name: 'Update scroll positions' }).click()
     await expect(page.getByText('Document scroll position is 0 & 0')).toBeVisible()
@@ -468,7 +463,8 @@ test.describe('preserve scroll', () => {
     consoleMessages.listen(page)
 
     await page.getByRole('link', { exact: true, name: 'Reset Scroll (Callback)' }).click()
-    await expect(page).toHaveURL('/links/preserve-scroll-false-page-two')
+    await expect(page).toHaveURL('/links/preserve-scroll-false-page-two?foo=foo')
+
     await expect(page.getByText('Foo is now foo')).toBeVisible()
 
     await ensureScrollPosition(page, '#slot', 10, 15)
@@ -489,7 +485,7 @@ test.describe('preserve scroll', () => {
   test('does not restore untracked scroll regions when pressing the back button', async ({ page }) => {
     await page.getByRole('link', { exact: true, name: 'Reset Scroll' }).click()
 
-    await expect(page).toHaveURL('/links/preserve-scroll-false-page-two')
+    await expect(page).toHaveURL('/links/preserve-scroll-false-page-two?foo=bar')
     await expect(page.getByText('Foo is now bar')).toBeVisible()
 
     await scrollElementTo(
@@ -519,7 +515,7 @@ test.describe('preserve scroll', () => {
     consoleMessages.listen(page)
     await page.getByRole('link', { exact: true, name: 'Preserve Scroll (Callback)' }).click()
 
-    await expect(page).toHaveURL('/links/preserve-scroll-false-page-two')
+    await expect(page).toHaveURL('/links/preserve-scroll-false-page-two?foo=baz')
     await expect(page.getByText('Foo is now baz')).toBeVisible()
 
     await scrollElementTo(
@@ -591,7 +587,7 @@ test.describe('enabled', () => {
 
   test('resets scroll regions to the top when doing a regular visit', async ({ page }) => {
     await page.getByText('Reset Scroll', { exact: true }).click()
-    await expect(page).toHaveURL('/links/preserve-scroll-page-two')
+    await expect(page).toHaveURL('/links/preserve-scroll-page-two?foo=bar')
     await expect(page.getByText('Foo is now bar')).toBeVisible()
     await page.getByRole('button', { exact: true, name: 'Update scroll positions' }).click()
     await expect(page.getByText('Document scroll position is 0 & 0')).toBeVisible()
@@ -602,7 +598,7 @@ test.describe('enabled', () => {
     consoleMessages.listen(page)
     await page.getByText('Reset Scroll (Callback)', { exact: true }).click()
 
-    await expect(page).toHaveURL('/links/preserve-scroll-page-two')
+    await expect(page).toHaveURL('/links/preserve-scroll-page-two?foo=foo')
     await expect(page.getByText('Foo is now foo')).toBeVisible()
     await page.getByRole('button', { exact: true, name: 'Update scroll positions' }).click()
     await expect(page.getByText('Document scroll position is 0 & 0')).toBeVisible()
@@ -620,7 +616,7 @@ test.describe('enabled', () => {
   test('preserves scroll regions when using the "preserve-scroll" feature', async ({ page }) => {
     await page.getByText('Preserve Scroll', { exact: true }).click()
 
-    await expect(page).toHaveURL('/links/preserve-scroll-page-two')
+    await expect(page).toHaveURL('/links/preserve-scroll-page-two?foo=baz')
     await expect(page.getByText('Foo is now baz')).toBeVisible()
     await page.getByRole('button', { exact: true, name: 'Update scroll positions' }).click()
     await expect(page.getByText('Document scroll position is 5 & 7')).toBeVisible()
@@ -631,7 +627,7 @@ test.describe('enabled', () => {
     consoleMessages.listen(page)
     await page.getByText('Preserve Scroll (Callback)', { exact: true }).click()
 
-    await expect(page).toHaveURL('/links/preserve-scroll-page-two')
+    await expect(page).toHaveURL('/links/preserve-scroll-page-two?foo=baz')
     await expect(page.getByText('Foo is now baz')).toBeVisible()
     await ensureScrollPosition(page, '#slot', 10, 15)
     await page.getByRole('button', { exact: true, name: 'Update scroll positions' }).click()
@@ -648,7 +644,7 @@ test.describe('enabled', () => {
 
   test('restores all tracked scroll regions when pressing the back button', async ({ page }) => {
     await page.getByTestId('preserve').click()
-    await expect(page).toHaveURL('/links/preserve-scroll-page-two')
+    await expect(page).toHaveURL('/links/preserve-scroll-page-two?foo=baz')
     await page.waitForTimeout(100)
 
     await scrollElementTo(
@@ -921,7 +917,7 @@ test.describe('partial reloads', () => {
     requests.listen(page)
 
     await page.getByRole('link', { name: 'Update All' }).click()
-    await expect(page).toHaveURL('/links/partial-reloads')
+    await expect(page).toHaveURL('/links/partial-reloads?foo=1')
 
     await expect(requests.requests).toHaveLength(1)
 
@@ -935,7 +931,7 @@ test.describe('partial reloads', () => {
     requests.listen(page)
 
     await page.getByRole('link', { name: 'Only foo + bar' }).click()
-    await expect(page).toHaveURL('/links/partial-reloads')
+    await expect(page).toHaveURL('/links/partial-reloads?foo=1')
 
     await expect(requests.requests).toHaveLength(1)
 
@@ -950,7 +946,7 @@ test.describe('partial reloads', () => {
 
   test('it updates all props when the feature is not being used', async ({ page }) => {
     await page.getByRole('link', { name: 'Update All' }).click()
-    await expect(page).toHaveURL('/links/partial-reloads')
+    await expect(page).toHaveURL('/links/partial-reloads?foo=1')
 
     await expect(page.getByText('Foo is now 2')).toBeVisible()
     await expect(page.getByText('Bar is now 3')).toBeVisible()
@@ -959,21 +955,21 @@ test.describe('partial reloads', () => {
 
   test('it only updates props that are passed through "only"', async ({ page }) => {
     await page.getByRole('link', { name: 'Only foo + bar' }).click()
-    await expect(page).toHaveURL('/links/partial-reloads')
+    await expect(page).toHaveURL('/links/partial-reloads?foo=1')
 
     await expect(page.getByText('Foo is now 2')).toBeVisible()
     await expect(page.getByText('Bar is now 3')).toBeVisible()
     await expect(page.getByText('Baz is now 3')).toBeVisible()
 
     await page.getByRole('link', { name: 'Only baz' }).click()
-    await expect(page).toHaveURL('/links/partial-reloads')
+    await expect(page).toHaveURL('/links/partial-reloads?foo=2')
 
     await expect(page.getByText('Foo is now 2')).toBeVisible()
     await expect(page.getByText('Bar is now 3')).toBeVisible()
     await expect(page.getByText('Baz is now 5')).toBeVisible()
 
     await page.getByRole('link', { name: 'Update All' }).click()
-    await expect(page).toHaveURL('/links/partial-reloads')
+    await expect(page).toHaveURL('/links/partial-reloads?foo=2')
 
     await expect(page.getByText('Foo is now 3')).toBeVisible()
     await expect(page.getByText('Bar is now 4')).toBeVisible()
@@ -982,14 +978,14 @@ test.describe('partial reloads', () => {
 
   test('it only updates props that are not passed through "except"', async ({ page }) => {
     await page.getByRole('link', { name: 'Except foo + bar' }).click()
-    await expect(page).toHaveURL('/links/partial-reloads')
+    await expect(page).toHaveURL('/links/partial-reloads?foo=1')
 
     await expect(page.getByText('Foo is now 1')).toBeVisible()
     await expect(page.getByText('Bar is now 2')).toBeVisible()
     await expect(page.getByText('Baz is now 4')).toBeVisible()
 
     await page.getByRole('link', { name: 'Except baz' }).click()
-    await expect(page).toHaveURL('/links/partial-reloads')
+    await expect(page).toHaveURL('/links/partial-reloads?foo=1')
 
     await expect(page.getByText('Foo is now 2')).toBeVisible()
     await expect(page.getByText('Bar is now 3')).toBeVisible()